Refactor: rename fiat payment methods to payment methods
